Using static const instead of #define in Constants.h

+++ GUI/Metal/TextureRect.swift
@@ -23,7 +23,7 @@ extension Canvas {
     // Returns the entire texture area (including HBLANK and VBLANK)
     var entire: CGRect {
         
-        return CGRect(x: 0, y: 0, width: Int(4 * TPP * HPOS_CNT), height: Int(VPOS_CNT))
+        return CGRect(x: 0, y: 0, width: 4 * Int(TPP) * HPOS_CNT, height: VPOS_CNT)
     }
     
     var entireNormalized: CGRect {
@@ -36,10 +36,10 @@ extension Canvas {
 
         let pal = amiga.agnus.isPAL
         
-        let x1 = Int(4 * TPP * HBLANK_CNT)
-        let x2 = Int(4 * TPP * HPOS_CNT_PAL)
-        let y1 = Int(VBLANK_CNT)
-        let y2 = pal ? Int(VPOS_CNT_PAL) : Int(VPOS_CNT_NTSC)
+        let x1 = 4 * Int(TPP) * HBLANK_CNT
+        let x2 = 4 * Int(TPP) * HPOS_CNT_PAL
+        let y1 = VBLANK_CNT
+        let y2 = pal ? VPOS_CNT_PAL : VPOS_CNT_NTSC
         
         return CGRect(x: x1, y: y1, width: x2 - x1, height: y2 - y1)
     }
@@ -147,8 +147,8 @@ extension Canvas {
         debug(.metal, "(\(x1),\(y1)) - \(x2),\(y2))")
 
         // Compensate the texture shift
-        x1 -= CGFloat(HBLANK_MIN * TPP) * 4
-        x2 -= CGFloat(HBLANK_MIN * TPP) * 4
+        x1 -= CGFloat(HBLANK_MIN) * CGFloat(TPP) * 4
+        x2 -= CGFloat(HBLANK_MIN) * CGFloat(TPP) * 4
         
         updateTextureRect()
     }
